Less code
